% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.hap.R
\name{perm.hap}
\alias{perm.hap}
\title{Permute haplotypes}
\usage{
perm.hap(haplotypes, preplist, generation.id)
}
\arguments{
\item{haplotypes}{a matrix of haplotypes to be permuted}

\item{preplist}{a list as output by \code{\link{Preparation.hap.perm}}}

\item{generation.id}{a vector of length(sample size) which
indicates if the subject is founder (\code{generation.id=0}),
a child from first generation (\code{generation.id=1}), a
child from second generation (\code{generation.id=2}), etc.
This vector can be calculated by the
\code{kinship2::kindepth()} function.}
}
\value{
a matrix of permuted haplotypes
}
\description{
Permute haplotypes
}
\author{
Karim Oualkacha

M'Hamed Lajmi Lakhal-Chaieb
}
\keyword{internal}

