% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preparation.VCC.R
\name{Preparation.VCC}
\alias{Preparation.VCC}
\title{Prepare for the VC-C methods}
\usage{
Preparation.VCC(n, RH0, Phi)
}
\arguments{
\item{n}{sample size (number of individuals)}

\item{RH0}{a list of the results obtained from the function
\code{\link{Estim.H0.VCC}}}

\item{Phi}{kinship matrix}
}
\value{
XXX
}
\description{
Prepare for the VC-C methods
}
\author{
Karim Oualkacha

M'Hamed Lajmi Lakhal-Chaieb
}
\keyword{internal}

