% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.VCC1.R
\name{pvalue.VCC1}
\alias{pvalue.VCC1}
\title{Compute score and p-value for the VC-C1 method}
\usage{
pvalue.VCC1(Prep, W, G)
}
\arguments{
\item{Prep}{a list of results obtained from the function
\code{\link{Preparation.VCC}}}

\item{G}{genotype matrix}
}
\value{
a list with the score and p-value of the association test
    in the given region. The list contains the elements
    \code{score} and \code{p.value}.
}
\description{
Compute score and p-value for the VC-C1 method
}
\author{
Karim Oualkacha

M'Hamed Lajmi Lakhal-Chaieb
}
\seealso{
\code{\link{pvalue.VCC2}}, \code{\link{pvalue.VCC3}}
}
\keyword{internal}

